tax_service/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.10
python-json-logger==2.0.4